moved google maps api information to location js file
